bug fixing
git-svn-id: https://svn.code.sf.net/p/openuss/code/trunk@2729 60046477-0c10-0410-bf4b-e552a4ff1768

+++ openuss-plexus/plexus/plexus-core/src/main/java/org/openuss/lecture/LectureIndex.java
@@ -77,15 +77,17 @@ public class LectureIndex extends DomainIndexer {
 	}
 
 	private void deleteIndex() {
-		String[] files = fsDirectory.list();
-		for (String file : files) {
-			Lock lock = fsDirectory.makeLock(file);
-			try {
-				fsDirectory.deleteFile(file);
-			} catch (IOException e) {
-				logger.error(e);
-			} finally {
-				lock.release();
+		if (fsDirectory != null) {
+			String[] files = fsDirectory.list();
+			for (String file : files) {
+				Lock lock = fsDirectory.makeLock(file);
+				try {
+					fsDirectory.deleteFile(file);
+				} catch (IOException e) {
+					logger.error(e);
+				} finally {
+					lock.release();
+				}
 			}
 		}
 	}
